Set the _DEBUG flag for CMake DEBUG builds (#113)

+++ Emulator/TMemory.cpp
@@ -690,7 +690,7 @@ Boolean
 TMemory::Read( VAddr inAddress, KUInt32& outWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&1)) {
 			KPrintf("Watchpoint 0x%08X read around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -730,7 +730,7 @@ Boolean
 TMemory::ReadAligned( VAddr inAddress, KUInt32& outWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&1)) {
 			KPrintf("Watchpoint 0x%08X read around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -770,7 +770,7 @@ inline Boolean
 TMemory::ReadROMRAM( VAddr inAddress, KUInt32& outWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&1)) {
 			KPrintf("Watchpoint 0x%08X read around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -1332,7 +1332,7 @@ Boolean
 TMemory::ReadB( VAddr inAddress, KUInt8& outByte )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&1)) {
 			KPrintf("Watchpoint 0x%08X read around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -1567,7 +1567,7 @@ Boolean
 TMemory::Write( VAddr inAddress, KUInt32 inWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&2)) {
 			KPrintf("Watchpoint 0x%08X written around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -1604,7 +1604,7 @@ Boolean
 TMemory::WriteAligned( VAddr inAddress, KUInt32 inWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&2)) {
 			KPrintf("Watchpoint 0x%08X written around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
@@ -1641,7 +1641,7 @@ inline Boolean
 TMemory::WriteRAM( VAddr inAddress, KUInt32 inWord )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&2)) {
 			KPrintf("Watchpoint 0x%08lX written around 0x%08lX\n", (unsigned long)inAddress, (unsigned long)mProcessor->mCurrentRegisters[15]);
@@ -2155,7 +2155,7 @@ Boolean
 TMemory::WriteB( VAddr inAddress, KUInt8 inByte )
 {
 #ifdef _DEBUG
-	int i;
+	size_t i;
 	for (i=0; i<mWPCount; i++) {
 		if ((mWatchpoints[i].fAddress==inAddress) && (mWatchpoints[i].fMode&2)) {
 			KPrintf("Watchpoint 0x%08X written around 0x%08X\n", (unsigned int)inAddress, (unsigned int)mProcessor->mCurrentRegisters[15]);
